requirements.txt
infi.systray==0.1.12
screen_brightness_control==0.11.5
